refactor: split auth_test.go to match source file layout
Split the 949-line auth_test.go along the Phase 4 source split:
token cache persistence tests to token_store_test.go, interactive
login flow tests to login_flow_test.go, payload parsing tests to
payload_test.go. Pure movement within the package; test count
verified unchanged (29 top-level tests).
